+++ src/Emulator/Cores/RiscV/Ri5cy.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2024 Antmicro
+// Copyright (c) 2010-2025 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -21,13 +21,13 @@ namespace Antmicro.Renode.Peripherals.CPU
 
             // register custom CSRs
             // TODO: add support for HW loops
-            RegisterCSR((ulong)0x7b0, () => 0u, _ => {}); //lpstart0
-            RegisterCSR((ulong)0x7b1, () => 0u, _ => {}); //lpend1
-            RegisterCSR((ulong)0x7b2, () => 0u, _ => {}); //lpcount0
+            RegisterCSR((ushort)0x7b0, () => 0u, _ => {}); //lpstart0
+            RegisterCSR((ushort)0x7b1, () => 0u, _ => {}); //lpend1
+            RegisterCSR((ushort)0x7b2, () => 0u, _ => {}); //lpcount0
 
-            RegisterCSR((ulong)0x7b4, () => 0u, _ => {}); //lpstart1
-            RegisterCSR((ulong)0x7b5, () => 0u, _ => {}); //lpend1
-            RegisterCSR((ulong)0x7b6, () => 0u, _ => {}); //lpcount1
+            RegisterCSR((ushort)0x7b4, () => 0u, _ => {}); //lpstart1
+            RegisterCSR((ushort)0x7b5, () => 0u, _ => {}); //lpend1
+            RegisterCSR((ushort)0x7b6, () => 0u, _ => {}); //lpcount1
         }
     }
 }
